Add FindingsFilter
