--- export-config/resource-role.yaml
@@ -1,35 +1,0 @@
-AWSTemplateFormatVersion: "2010-09-09"
-Description: >
-  This CloudFormation template creates a role assumed by CloudFormation
-  during CRUDL operations to mutate resources on behalf of the customer.
-
-Resources:
-  ExecutionRole:
-    Type: AWS::IAM::Role
-    Properties:
-      MaxSessionDuration: 8400
-      AssumeRolePolicyDocument:
-        Version: '2012-10-17'
-        Statement:
-          - Effect: Allow
-            Principal:
-              Service: resources.cloudformation.amazonaws.com
-            Action: sts:AssumeRole
-      Path: "/"
-      Policies:
-        - PolicyName: ResourceTypePolicy
-          PolicyDocument:
-            Version: '2012-10-17'
-            Statement:
-              - Effect: Allow
-                Action:
-                - "initech:CreateReport"
-                - "initech:DeleteReport"
-                - "initech:DescribeReport"
-                - "initech:ListReports"
-                - "initech:UpdateReport"
-                Resource: "*"
-Outputs:
-  ExecutionRoleArn:
-    Value:
-      Fn::GetAtt: ExecutionRole.Arn
