--- export-config/src/main/java/software/amazon/macie/classificationexportconfiguration/CallbackContext.java
@@ -1,12 +1,0 @@
-package software.amazon.macie.classificationexportconfiguration;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@Builder
-public class CallbackContext {
-
-}
