--- export-config/src/main/java/software/amazon/macie/classificationexportconfiguration/Configuration.java
@@ -1,12 +1,0 @@
-package software.amazon.macie.classificationexportconfiguration;
-
-import java.util.Map;
-import org.json.JSONObject;
-import org.json.JSONTokener;
-
-class Configuration extends BaseConfiguration {
-
-    public Configuration() {
-        super("aws-macie-classificationexportconfiguration.json");
-    }
-}
