+++ findings-filter/lombok.config
