+++ findings-filter/src/main/java/software/amazon/macie/findingsfilter/CallbackContext.java
@@ -1,0 +1,12 @@
+package software.amazon.macie.findingsfilter;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@Builder
+public class CallbackContext {
+
+}
