+++ findings-filter/src/main/java/software/amazon/macie/findingsfilter/Configuration.java
@@ -1,0 +1,12 @@
+package software.amazon.macie.findingsfilter;
+
+import java.util.Map;
+import org.json.JSONObject;
+import org.json.JSONTokener;
+
+class Configuration extends BaseConfiguration {
+
+    public Configuration() {
+        super("aws-macie-findingsfilter.json");
+    }
+}
