+++ findings-filter/src/main/java/software/amazon/macie/findingsfilter/ListHandler.java
@@ -1,0 +1,30 @@
+package software.amazon.macie.findingsfilter;
+
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ListHandler extends BaseHandler<CallbackContext> {
+
+    @Override
+    public ProgressEvent<ResourceModel, CallbackContext> handleRequest(
+        final AmazonWebServicesClientProxy proxy,
+        final ResourceHandlerRequest<ResourceModel> request,
+        final CallbackContext callbackContext,
+        final Logger logger) {
+
+        final List<ResourceModel> models = new ArrayList<>();
+
+        // TODO : put your code here
+
+        return ProgressEvent.<ResourceModel, CallbackContext>builder()
+            .resourceModels(models)
+            .status(OperationStatus.SUCCESS)
+            .build();
+    }
+}
