+++ findings-filter/template.yml
@@ -1,0 +1,22 @@
+AWSTemplateFormatVersion: "2010-09-09"
+Transform: AWS::Serverless-2016-10-31
+Description: AWS SAM template for the AWS::Macie::FindingsFilter resource type
+
+Globals:
+  Function:
+    Timeout: 60  # docker start-up times can be long for SAM CLI
+
+Resources:
+  TypeFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: software.amazon.macie.findingsfilter.HandlerWrapper::handleRequest
+      Runtime: java8
+      CodeUri: ./target/aws-macie-findingsfilter-handler-1.0-SNAPSHOT.jar
+
+  TestEntrypoint:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: software.amazon.macie.findingsfilter.HandlerWrapper::testEntrypoint
+      Runtime: java8
+      CodeUri: ./target/aws-macie-findingsfilter-handler-1.0-SNAPSHOT.jar
